requirements.txt

configparser
colorama


# Google models
google-genai
google-generativeai

# OpenAI models
openai
